Split langtool into multiple source file.
